adding correct components to table cell in user diary page

+++ src/frontend/src/utils/general.ts
@@ -2,7 +2,7 @@ export const getMonthName = (monthNumber: number) => {
     const date = new Date();
     date.setMonth(monthNumber - 1);
 
-    return date.toLocaleString('en-US', { month: 'long' });
+    return date.toLocaleString('en-US', { month: 'short' });
 };
 
 export const kFormatter = (num: number) => {
